that's all folks

+++ plugins/purge.py
@@ -10,6 +10,8 @@ def purge(client: Client, message: Message):
     messages = client.iter_history(message.chat.id, offset_id=message.reply_to_message.message_id, reverse=True)
     for message in messages:
         try:
+            if 'me' in message.command and not message.from_user.is_self:
+                continue
             client.delete_messages(message.chat.id, message.message_id)
         except FloodWait as e:
             sleep(e.x)
